update script to work with updated annotid table
Annotids are stored on a VM in a database table.
The ones in that table got out of sync with the ones on pn-opus which sometimes resulted in duplicated annotids.
So, I recreated the table.
But this time it only contains annotids and no other information.
So, the scripts had to be update a bit.